+++ baybe/recommenders/meta/base.py
@@ -27,6 +27,7 @@ class MetaRecommender(SerialMixin, RecommenderProtocol, ABC):
         searchspace: SearchSpace,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> PureRecommender:
         """Select a pure recommender for the given experimentation context.
 
@@ -39,6 +40,8 @@ class MetaRecommender(SerialMixin, RecommenderProtocol, ABC):
                 See :func:`baybe.recommenders.meta.base.MetaRecommender.recommend`.
             measurements:
                 See :func:`baybe.recommenders.meta.base.MetaRecommender.recommend`.
+            pending_experiments:
+                See :func:`baybe.recommenders.meta.base.MetaRecommender.recommend`.
 
         Returns:
             The selected recommender.
@@ -50,6 +53,7 @@ class MetaRecommender(SerialMixin, RecommenderProtocol, ABC):
         searchspace: SearchSpace,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> pd.DataFrame:
         """See :func:`baybe.recommenders.base.RecommenderProtocol.recommend`."""
         recommender = self.select_recommender(
@@ -57,6 +61,7 @@ class MetaRecommender(SerialMixin, RecommenderProtocol, ABC):
             searchspace=searchspace,
             objective=objective,
             measurements=measurements,
+            pending_experiments=pending_experiments,
         )
 
         # Non-predictive recommenders should not be called with an objective or
@@ -72,7 +77,10 @@ class MetaRecommender(SerialMixin, RecommenderProtocol, ABC):
         )
 
         return recommender.recommend(
-            batch_size=batch_size, searchspace=searchspace, **optional_args
+            batch_size=batch_size,
+            searchspace=searchspace,
+            pending_experiments=pending_experiments,
+            **optional_args,
         )