+++ baybe/recommenders/meta/sequential.py
@@ -56,6 +56,7 @@ class TwoPhaseMetaRecommender(MetaRecommender):
         searchspace: SearchSpace | None = None,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> PureRecommender:
         # See base class.
 
@@ -135,6 +136,7 @@ class SequentialMetaRecommender(MetaRecommender):
         searchspace: SearchSpace | None = None,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> PureRecommender:
         # See base class.
 
@@ -224,6 +226,7 @@ class StreamingSequentialMetaRecommender(MetaRecommender):
         searchspace: SearchSpace | None = None,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> PureRecommender:
         # See base class.