+++ baybe/recommenders/pure/nonpredictive/base.py
@@ -4,12 +4,13 @@ import warnings
 from abc import ABC
 
 import pandas as pd
+from attr import fields
 from attrs import define
 
 from baybe.exceptions import UnusedObjectWarning
 from baybe.objectives.base import Objective
 from baybe.recommenders.pure.base import PureRecommender
-from baybe.searchspace.core import SearchSpace
+from baybe.searchspace.core import SearchSpace, SearchSpaceType
 
 
 @define
@@ -22,6 +23,7 @@ class NonPredictiveRecommender(PureRecommender, ABC):
         searchspace: SearchSpace,
         objective: Objective | None = None,
         measurements: pd.DataFrame | None = None,
+        pending_experiments: pd.DataFrame | None = None,
     ) -> pd.DataFrame:
         # See base class.
 
@@ -39,9 +41,22 @@ class NonPredictiveRecommender(PureRecommender, ABC):
                 f"consider any objectives, meaning that the argument is ignored.",
                 UnusedObjectWarning,
             )
+        if (pending_experiments is not None) and (
+            self.allow_recommending_pending_experiments
+            or searchspace.type is not SearchSpaceType.DISCRETE
+        ):
+            warnings.warn(
+                f"Pending experiments were provided but the selected recommender "
+                f"'{self.__class__.__name__}' only utilizes this information for "
+                f"purely discrete spaces and "
+                f"{fields(self.__class__).allow_recommending_pending_experiments.name}"
+                f"=False.",
+                UnusedObjectWarning,
+            )
         return super().recommend(
             batch_size=batch_size,
             searchspace=searchspace,
             objective=objective,
             measurements=measurements,
+            pending_experiments=pending_experiments,
         )